Move interiors and assets into new folders
