+++ assets/visual/interior designs/observation_windows.afdesign
