+++ assets/visual/interior designs/stellar_cafe.afdesign
