+++ assets/visual/observation_windows.png
