+++ assets/visual/stellar_cafe.png
